Test files added and notes to code added
Added E. coli test cases. This is a small genome so it should be a little easier to work with for now. We will move to GRCm39 later though. Also added some notes that we will have to handle cases where read lengths are not all equal. Minor changes to the genome size algorithm were also made.